moved CircularPercentageIndicator to engine, add unit tests
