+++ android/engine/src/main/res/values/strings.xml
@@ -107,4 +107,5 @@
     <string name="pin_menu_login">Login</string>
     <string name="forgot_pin">Forgot pin?</string>
     <string name="please_contact_supervisor">Please contact your supervisor.</string>
+    <string name="percentage">%</string>
 </resources>
